api/src/main/resources/requirements.txt
argparse==1.2.1
certifi==2017.4.17
cffi==1.9.1
cryptography==1.7.1
enum34==1.1.6
happybase==1.0.0
idna==2.1
ipaddress==1.0.17
ordereddict==1.1
paramiko==2.1.1
ply==3.9
pyasn1==0.1.9
pycparser==2.17
pyhs2==0.6.0
pywebhdfs==0.4.1
PyYAML==3.12
requests==2.12.4
sasl==0.2.1
singledispatch==3.4.0.3
six==1.10.0
spur==0.3.20
starbase==0.3.3
thrift==0.9.3
thriftpy==0.3.9
tornado==4.4.2
tornado-cors==0.6.0
